test(herdr): move tests into test/ so CI picks them up [no-feature-entry]
CI runs `bun test ./test` only; these 5 files were mistakenly created
under src/ so they never ran. git mv into test/, imports rewritten to
../src/*. src/herdr.test.ts collided with a pre-existing
test/herdr.test.ts, so it's renamed to test/herdr-listasync.test.ts
(mirrors poller-listasync.test.ts; both cover the new listAsync surface).
